requirements.txt
pact-python==1.7.0
pytest==7.2.2